src/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
aiohttp==3.9.5
    # via fastapi_app (pyproject.toml)
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
async-timeout==4.0.3
    # via asyncpg
asyncpg==0.29.0
    # via fastapi_app (pyproject.toml)
attrs==23.2.0
    # via aiohttp
azure-core==1.30.1
    # via azure-identity
azure-identity==1.16.0
    # via fastapi_app (pyproject.toml)
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via uvicorn
cryptography==42.0.6
    # via
    #   azure-identity
    #   msal
    #   pyjwt
distro==1.9.0
    # via openai
environs==11.0.0
    # via fastapi_app (pyproject.toml)
fastapi==0.110.3
    # via fastapi_app (pyproject.toml)
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.0.3
    # via sqlalchemy
gunicorn==20.1.0
    # via fastapi_app (pyproject.toml)
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via openai
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
markdown-it-py==3.0.0
    # via rich
marshmallow==3.21.2
    # via environs
mdurl==0.1.2
    # via markdown-it-py
msal==1.28.0
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.1.0
    # via azure-identity
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
numpy==1.26.4
    # via pgvector
openai==1.25.2
    # via
    #   fastapi_app (pyproject.toml)
    #   openai-messages-token-helper
openai-messages-token-helper==0.1.3
    # via fastapi_app (pyproject.toml)
packaging==24.0
    # via
    #   marshmallow
    #   msal-extensions
pgvector==0.2.5
    # via fastapi_app (pyproject.toml)
pillow==10.3.0
    # via openai-messages-token-helper
portalocker==2.8.2
    # via msal-extensions
pycparser==2.22
    # via cffi
pydantic==2.7.1
    # via
    #   fastapi
    #   openai
pydantic-core==2.18.2
    # via pydantic
pygments==2.18.0
    # via rich
pyjwt[crypto]==2.8.0
    # via
    #   msal
    #   pyjwt
python-dotenv==1.0.1
    # via
    #   environs
    #   fastapi_app (pyproject.toml)
    #   uvicorn
pyyaml==6.0.1
    # via uvicorn
regex==2024.4.28
    # via tiktoken
requests==2.31.0
    # via
    #   azure-core
    #   msal
    #   tiktoken
rich==13.7.1
    # via fastapi_app (pyproject.toml)
six==1.16.0
    # via azure-core
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
sqlalchemy[asyncio]==2.0.30
    # via fastapi_app (pyproject.toml)
starlette==0.37.2
    # via fastapi
tiktoken==0.6.0
    # via
    #   fastapi_app (pyproject.toml)
    #   openai-messages-token-helper
tqdm==4.66.4
    # via openai
typing-extensions==4.11.0
    # via
    #   azure-core
    #   fastapi
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
urllib3==2.2.1
    # via requests
uvicorn[standard]==0.23.2
    # via fastapi_app (pyproject.toml)
uvloop==0.19.0
    # via uvicorn
watchfiles==0.21.0
    # via uvicorn
websockets==12.0
    # via uvicorn
yarl==1.9.4
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
